fapi: cleanup callbacks
The callbacks were implmented in a fashion where their was 10 versions
of each callback function in C and each version would get used in a
set_<fn>_callback routine. Ie the first to set would get assigned
the C callback ending in 0, the next caller would get index 1, and
so forth until the index's were exhausted. This placed in artificial
limitation on the number of callback users. Instead, have a single
C function callback for ALL FAPI context callers and demux based on the
callable and userdata.

SetCallback functions associate an _FAPI_CB_UDATA object which contains:
1. The Python Callback Callable
2. The Userdata
3. Any exceptions from the Callable that need to be re-raised.

This is all stored in the userdata for the C callback. If None is passed
for the callback, the state is cleared and the C callback is set to NULL.

A Fapi Function may then invoke a callback, if set, the userdata will be
the _FAPI_CB_UDATA, which we can then use to get to the user defined
callback and userdata. We can also use this object to get exceptions in
the python callback back accross the C interface and re-raise them if
needed.

TODO:
- Test policy action callback unset
- Test set_sign_callback unset
- Test set_branch_callback unset

+++ test/test_fapi.py
@@ -16,6 +16,7 @@ from tpm2_pytss import *
 from tpm2_pytss.internal.utils import is_bug_fixed
 
 from .TSS2_BaseTest import TpmSimulator
+from tpm2_pytss.TSS2_Exception import TSS2_Exception
 
 pytestmark = pytest.mark.skipif(
     "tpm2_pytss.FAPI" not in sys.modules, reason="FAPI Not Detected"
@@ -781,6 +782,12 @@ class Common:
         # use key for signing: success
         self.fapi.sign(path=key_path, digest=b"\x11" * 32)
 
+        # unset signing callback, should fail
+        self.fapi.set_sign_callback(callback=None)
+
+        with pytest.raises(TSS2_Exception):
+            self.fapi.sign(path=key_path, digest=b"\x11" * 32)
+
     def test_policy_branched(self):
         pcr_index = 15
         pcr_data = b"ABCDEF"
@@ -885,6 +892,15 @@ class Common:
         nv_data, _ = self.fapi.nv_read(nv_path)
         assert nv_data == b"Hello World"
 
+        # test without a callback set
+        if not is_bug_fixed(fixed_in="3.1", backports=["2.4.3", "3.0.1", "3.1.0"]):
+            # skip tests as theirs a bug
+            return
+
+        self.fapi.set_branch_callback(callback=None)
+        with pytest.raises(TSS2_Exception):
+            self.fapi.nv_read(nv_path)
+
         # clean up
         self.fapi.delete(path=nv_path)
 
@@ -938,6 +954,12 @@ class Common:
 
         # use key for signing: success
         self.fapi.sign(path=key_path, digest=b"\x11" * 32)
+
+        # set policy Action callback to NULL
+        self.fapi.set_policy_action_callback(callback=None)
+
+        with pytest.raises(TSS2_Exception):
+            self.fapi.sign(path=key_path, digest=b"\x11" * 32)
 
 
 @pytest.mark.usefixtures("init_fapi_ecc")
